+++ src/App/backend/test/Altinn.App.Api.Tests/Mocks/ProcessClientMock.cs
@@ -50,7 +50,11 @@ public class ProcessClientMock : IProcessClient
         }
     }
 
-    public Task<ProcessHistoryList> GetProcessHistory(string instanceGuid, string instanceOwnerPartyId)
+    public Task<ProcessHistoryList> GetProcessHistory(
+        string instanceGuid,
+        string instanceOwnerPartyId,
+        StorageAuthenticationMethod? authenticationMethod = null
+    )
     {
         throw new NotImplementedException();
     }