+++ src/Designer/frontend/app-development/hooks/mutations/useAddLayoutSetMutation.test.ts
@@ -2,17 +2,16 @@ import { queriesMock } from 'app-shared/mocks/queriesMock';
 import { renderHookWithProviders } from '../../test/mocks';
 import { waitFor } from '@testing-library/react';
 import { useAddLayoutSetMutation } from './useAddLayoutSetMutation';
-import type { LayoutSetConfig, LayoutSets } from 'app-shared/types/api/LayoutSetsResponse';
+import type { LayoutSetConfig } from 'app-shared/types/api/LayoutSetsResponse';
 import { createQueryClientMock } from 'app-shared/mocks/queryClientMock';
 import { QueryKey } from 'app-shared/types/QueryKey';
 import { app, org } from '@studio/testing/testids';
 
 // Test data:
-const layoutSetIdToUpdate = 'oldLayoutSetName';
 const taskType = 'data';
 const layoutSetConfig: LayoutSetConfig = {
   id: 'newLayoutSetName',
-  tasks: ['task_2'],
+  taskId: 'task_2',
 };
 
 describe('useAddLayoutSetMutation', () => {
@@ -23,7 +22,6 @@ describe('useAddLayoutSetMutation', () => {
       .renderHookResult.result;
     await waitFor(() =>
       addLayoutSetResult.current.mutateAsync({
-        layoutSetIdToUpdate,
         taskType,
         layoutSetConfig,
       }),
@@ -31,7 +29,7 @@ describe('useAddLayoutSetMutation', () => {
     expect(addLayoutSetResult.current.isSuccess).toBe(true);
 
     expect(queriesMock.addLayoutSet).toHaveBeenCalledTimes(1);
-    expect(queriesMock.addLayoutSet).toHaveBeenCalledWith(org, app, layoutSetIdToUpdate, {
+    expect(queriesMock.addLayoutSet).toHaveBeenCalledWith(org, app, {
       layoutSetConfig,
       taskType,
     });
@@ -42,81 +40,33 @@ describe('useAddLayoutSetMutation', () => {
       .renderHookResult.result;
     await waitFor(() =>
       addLayoutSetResult.current.mutateAsync({
-        layoutSetIdToUpdate,
         layoutSetConfig,
       }),
     );
     expect(addLayoutSetResult.current.isSuccess).toBe(true);
 
     expect(queriesMock.addLayoutSet).toHaveBeenCalledTimes(1);
-    expect(queriesMock.addLayoutSet).toHaveBeenCalledWith(org, app, layoutSetIdToUpdate, {
+    expect(queriesMock.addLayoutSet).toHaveBeenCalledWith(org, app, {
       layoutSetConfig,
-      undefined,
+      taskType: undefined,
     });
   });
 
-  it('Sets queryData if response is of type LayoutSets', async () => {
-    const layoutSets = { sets: [{ id: 'set1', tasks: ['Task_1'] }] };
+  it('Invalidates LayoutSetsExtended and LayoutSets caches on success', async () => {
     const queryClientMock = createQueryClientMock();
-    const addLayoutSetMock = jest
-      .fn()
-      .mockImplementation(() => Promise.resolve<LayoutSets>(layoutSets));
+    const invalidateQueriesSpy = jest.spyOn(queryClientMock, 'invalidateQueries');
     const addLayoutSetResult = renderHookWithProviders(
-      { addLayoutSet: addLayoutSetMock },
+      {},
       queryClientMock,
     )(() => useAddLayoutSetMutation(org, app)).renderHookResult.result;
-    await waitFor(() =>
-      addLayoutSetResult.current.mutateAsync({
-        layoutSetIdToUpdate,
-        taskType,
-        layoutSetConfig,
-      }),
-    );
+    await waitFor(() => addLayoutSetResult.current.mutateAsync({ taskType, layoutSetConfig }));
     expect(addLayoutSetResult.current.isSuccess).toBe(true);
 
-    const queryData = queryClientMock.getQueryData([QueryKey.LayoutSets, org, app]);
-    expect(queryData).toBe(layoutSets);
-  });
-
-  it('does not set queryData if response is not an object', async () => {
-    const addLayoutSetMock = jest
-      .fn()
-      .mockImplementation(() => Promise.resolve('Layout set already exist error'));
-    const queryClientMock = createQueryClientMock();
-    const addLayoutSetResult = renderHookWithProviders(
-      { addLayoutSet: addLayoutSetMock },
-      queryClientMock,
-    )(() => useAddLayoutSetMutation(org, app)).renderHookResult.result;
-    await waitFor(() =>
-      addLayoutSetResult.current.mutateAsync({
-        layoutSetIdToUpdate: layoutSetIdToUpdate,
-        taskType,
-        layoutSetConfig,
-      }),
-    );
-    expect(addLayoutSetResult.current.isSuccess).toBe(true);
-
-    const queryData = queryClientMock.getQueryData([QueryKey.LayoutSets, org, app]);
-    expect(queryData).toBe(undefined);
-  });
-
-  it('does not set queryData if response is not of type LayoutSets', async () => {
-    const addLayoutSetMock = jest.fn().mockImplementation(() => Promise.resolve({}));
-    const queryClientMock = createQueryClientMock();
-    const addLayoutSetResult = renderHookWithProviders(
-      { addLayoutSet: addLayoutSetMock },
-      queryClientMock,
-    )(() => useAddLayoutSetMutation(org, app)).renderHookResult.result;
-    await waitFor(() =>
-      addLayoutSetResult.current.mutateAsync({
-        layoutSetIdToUpdate: layoutSetIdToUpdate,
-        taskType,
-        layoutSetConfig,
-      }),
-    );
-    expect(addLayoutSetResult.current.isSuccess).toBe(true);
-
-    const queryData = queryClientMock.getQueryData([QueryKey.LayoutSets, org, app]);
-    expect(queryData).toBe(undefined);
+    expect(invalidateQueriesSpy).toHaveBeenCalledWith({
+      queryKey: [QueryKey.LayoutSetsExtended, org, app],
+    });
+    expect(invalidateQueriesSpy).toHaveBeenCalledWith({
+      queryKey: [QueryKey.LayoutSets, org, app],
+    });
   });
 });