+++ src/Designer/frontend/packages/process-editor/src/components/ConfigPanel/ConfigEndEvent/CustomReceiptContent/CreateCustomReceiptForm/CreateCustomReceiptForm.tsx
@@ -64,11 +64,10 @@ export const CreateCustomReceiptForm = ({
     const customReceiptLayoutSetConfig: LayoutSetConfig = {
       id: customReceipt.layoutSetId,
       dataType: customReceipt.dataModelId,
-      tasks: [PROTECTED_TASK_NAME_CUSTOM_RECEIPT],
+      taskId: PROTECTED_TASK_NAME_CUSTOM_RECEIPT,
     };
     addLayoutSet(
       {
-        layoutSetIdToUpdate: customReceipt.layoutSetId,
         layoutSetConfig: customReceiptLayoutSetConfig,
       },
       {